Additional UI changes, typescript definition updates.
